requirements/dev.txt
-r prod.txt

coverage>=4.3.4
coveralls>=1.1
flake8>=3.3.0
pytest>=3.0.7
pytest-django>=3.1.2
pytest-factoryboy>=1.3.1
pytest-mock>=1.5.0
pytest-pythonpath>=0.7.1
tox>=2.6.0
